+++ app/controllers/concepts_controller.rb
@@ -78,16 +78,6 @@ class ConceptsController < ApplicationController
     render nothing: true unless @query
   end
 
-  def info_external
-    @concept = Concept.new(short_name: params[:short_name], totalnum: params[:totalnum], key: params[:key], source_id: params[:source_id])
-    @query_concept = QueryConcept.new(external_key: params[:key], source_id: params[:source_id])
-    @query = current_user.all_queries.find_by_id(params[:query_id])
-
-    @mapping = nil
-
-    render 'info'
-  end
-
   def show
     @concept = Concept.find_by_id(params[:id])
     unless @concept