+++ app/models/report_concept.rb
@@ -20,15 +20,4 @@ class ReportConcept < ActiveRecord::Base
     self.attributes.reject{|key, val| ['id', 'report_id', 'created_at', 'updated_at'].include?(key.to_s)}
   end
 
-
-  def external_concept_information(current_user)
-    @external_concept_information ||= begin
-      information = {name: self.external_key}
-      if self.source
-        result_hash = self.source.external_concept_information(current_user, self.external_key)
-        information = result_hash[:result] if result_hash[:error].blank?
-      end
-      information
-    end
-  end
 end