+++ test/controllers/query_concepts_controller_test.rb
@@ -132,15 +132,6 @@ class QueryConceptsControllerTest < ActionController::TestCase
     assert_template 'query_concepts'
   end
 
-  test "should create query concept for external concept" do
-    assert_difference('QueryConcept.count') do
-      post :create, query_concept: @query_concept.attributes, external_key: 'external_key', source_id: sources(:three).to_param, query_id: queries(:one).to_param, format: 'js'
-    end
-
-    assert_not_nil assigns(:query)
-    assert_template 'query_concepts'
-  end
-
   test "should not create query concept without valid concept" do
     assert_difference('QueryConcept.count', 0) do
       post :create, query_concept: @query_concept.attributes, query_id: queries(:one).to_param, selected_concept_id: '-1', format: 'js'