+++ test/controllers/report_concepts_controller_test.rb
@@ -17,17 +17,6 @@ class ReportConceptsControllerTest < ActionController::TestCase
     assert_template 'report_concepts'
   end
 
-  test "should create report concept for external concept" do
-    assert_difference('ReportConcept.count') do
-      post :create, query_id: queries(:one).to_param, report_id: reports(:report).to_param, concept_id: "#{sources(:i2b2_valid).to_param},external_concept_key", format: 'js'
-    end
-
-    assert_not_nil assigns(:query)
-    assert_not_nil assigns(:report)
-    assert_equal (1..assigns(:report).report_concepts.size).to_a, assigns(:report).report_concepts.collect(&:position)
-    assert_template 'report_concepts'
-  end
-
   test "should not create report concept without valid report and query" do
     assert_difference('ReportConcept.count', 0) do
       post :create, query_id: -1, report_id: -1, concept_id: concepts(:continuous).to_param, format: 'js'